+++ src/utils.cpp
@@ -4,7 +4,7 @@
 
 uint64 GetCurrentTime()
 {
-    return std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
+    return (uint64)std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
 }
 
 uint32 GetTimeDiff(uint64 curTime, uint64 prevTime)
